wip: subagent 完了時の自動コミット (1 files)
PDCA kit の SubagentStop hook による自動コミットです。
PR 前に squash して意味のある単位にまとめてください。

M	src/rc_basics_lab/diagnostics/esp.py

PDCA-Kit-Auto-Commit: true

+++ src/rc_basics_lab/diagnostics/esp.py
@@ -337,7 +337,7 @@ def conditional_lyapunov(
             states, propagator, context.washout, last_start, cfg.propagator_tol
         )
 
-    scale = np.sqrt(n_units)
+    scale = float(np.sqrt(n_units))
     rng = np.random.default_rng(
         _DIRECTION_SEED if context.seed is None else context.seed
     )
